First commit
I had gone through Part 8 (User authentication and permissions) of the tutorial
and things were working pretty well.  I was tackling the challenge of creating
a page to show all books out on loan for use by librarians only but was having
trouble restriction the list view to only librarians.