Finished chapter 8 challenge
The challenge at the end of chapter 8 was "create a ... page that is only visible for librarians, that displays all books that have been borrowed, and which includes the name of each borrower."

I don't think I implemented it quite the way that the tutorial did in its final work.  I really wanted the check to be based simply on whether the user is in the `Librarians` group or not and I finally was able to do that.