Adding migration files for PythonAnywhere
I'm still having trouble setting up the site on PythonAnywhere.  I sent up a different tutorial a couple of weeks ago and was following similar steps.  The Mozilla tutorial is failing on the database migration site and I remembered that I had decided to not include the catalog/migrations/ files in the git repo.  When I revisit the repo for the other tutorial, I do see migrations files there.  I'm going to add the files and hope this will make PythonAnywhere happy.