Correcting login templates
When I tried to login to the site at PythonAnywhere, I got this nastygram:

  TemplateDoesNotExist at /accounts/login/

  registration/login.html
  Request Method:	GET
  Request URL:	http://pfuntner.pythonanywhere.com/accounts/login/?next=/catalog/

Logging still worked in the local PyCharm app so I wasn't sure why this was happening. I looked at the files in PyCharm and noticed that the templates/ directory that holds the login templates was under the base directory (locallibrary) instead of locallibrary/locallibrary.  I moved the templates directory in the local copy and login continued to work.  I will try pushing the changes to PythonAnywhere.